Update cloudshell_es_install_script.sh
fixes running the script with sudo. the issue was that when you  run the script with sudo the word python2.7 and python is searched at /usr/bin instead of /usr/bin/local. that's because when you run something with sudo the PATH environment variable lacks /usr/local/bin, in contrary when you run a script without sudo.